Squashed 'lodash-deep/' changes from ac17915..7aefcb8
7aefcb8 Revert "Revert "Fix versions""
b79cbe4 Revert "Fix versions"
f545aac added a defaultVal parameter to deepGet.  - the path to the value and the value itself is initialised if not existing  - if any path element can be interpreted as a numeric index and that element it is non-exiting an Array is created.  - in all other cases an object is created (in line with deepSet)
3bcac0b Merge commit '8ad586e5485b6e65288bedcadf2f7897f2c77d9a'
2646023 renamed mixins.deepStringifyPropertyPath to mixins.deepStringifyProperty to avoide double `string` in the name which makes it not more clear what the function does

git-subtree-dir: lodash-deep
git-subtree-split: 7aefcb8ebc223e8350e54689fc01910a488ba37d